refactor!: move ComponentDataGenerator under the renderer package
