+++ vaadin-renderer-flow-parent/vaadin-renderer-flow/src/main/java/com/vaadin/flow/data/renderer/ComponentDataGenerator.java
@@ -13,12 +13,12 @@
  * License for the specific language governing permissions and limitations under
  * the License.
  */
-package com.vaadin.flow.data.provider;
+package com.vaadin.flow.data.renderer;
 
 import org.slf4j.LoggerFactory;
 
 import com.vaadin.flow.component.Component;
-import com.vaadin.flow.data.renderer.ComponentRenderer;
+import com.vaadin.flow.data.provider.AbstractComponentDataGenerator;
 import com.vaadin.flow.dom.Element;
 import com.vaadin.flow.function.ValueProvider;
 
@@ -32,7 +32,7 @@ import elemental.json.JsonObject;
  * <p>
  * This class is used internally by listing components that support
  * ComponentRenderers.
- * 
+ *
  * @author Vaadin Ltd
  * @since 1.0.
  *
@@ -49,7 +49,7 @@ public class ComponentDataGenerator<T>
 
     /**
      * Creates a new generator.
-     * 
+     *
      * @param componentRenderer
      *            the renderer used to produce components based on data items
      * @param keyMapper
